+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.7.0] - 2026-04-10
+
+### Added
+
+- Add JMasterDetail component [#559](https://github.com/nbbrd/java-desktop-util/issues/559)
+- Add JRangeSlider component [#560](https://github.com/nbbrd/java-desktop-util/issues/560)
+
+### Fixed
+
+- Fix JGrid focus [#561](https://github.com/nbbrd/java-desktop-util/issues/561)
+
 ## [2.6.2] - 2026-04-09
 
 ### Fixed
@@ -147,7 +158,8 @@ java-desktop-util follows [semantic versioning](https://semver.org/).
 
 - JDK11 cleanup
 
-[Unreleased]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.2...HEAD
+[Unreleased]: https://github.com/nbbrd/java-desktop-util/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.2...v2.7.0
 [2.6.2]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.1...v2.6.2
 [2.6.1]: https://github.com/nbbrd/java-desktop-util/compare/v2.6.0...v2.6.1
 [2.6.0]: https://github.com/nbbrd/java-desktop-util/compare/v2.5.0...v2.6.0